Remove Fixity.hs-boot and move Fixity' to Syntax.Common
